lxc: fix lxc-auto init script
Commit 2cde10b95053 ("lxc: export systemd cgroups after install") adds
the lxc-auto init script to the lxc-auto package's postinst. This causes
the init script to be executed during package build, which in turn
causes the following build failure:

/home/stijn/Development/OpenWrt/openwrt/build_dir/target-x86_64_musl/root-x86/etc/init.d/lxc-auto: line 3: /lib/functions.sh: No such file or directory
postinst script ./usr/lib/opkg/info/lxc-auto.postinst has failed with exit code 1

Fix this by not sourcing /lib/functions.sh in the init script. It is not
needed as /etc/rc.common sources this anyway.

Fixes: 2cde10b95053 ("lxc: export systemd cgroups after install")
Suggested-by: Jo-Philipp Wich <jo@mein.io>
Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>